+++ src/include/S3/S3PutObjectRequest.h
@@ -83,7 +83,7 @@
 @property (nonatomic, retain) NSInputStream *stream;
 
 /** Number of milliseconds before expiration. */
-@property (nonatomic, assign, readonly) NSInteger expires;
+@property (nonatomic, assign, readonly) int32_t expires;
 
 /** Path of file in the filesystem to be uploaded.
  * The <code>contentLength</code> and <code>contentType</code> will be inferred.
